Remove NewsPostDrawableDate
